requirements.txt
sqlmodel==0.0.14
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.24.3
matplotlib==3.7.2
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
alembic==1.13.1
pydantic==2.5.0
pydantic-settings==2.1.0 